+++ electron-vite-project/src/Control.tsx
@@ -12,7 +12,7 @@
  *
  * @file Control.tsx
  */
-import React, { useContext, useEffect, useLayoutEffect, useState } from 'react';
+import React, { useCallback, useContext, useEffect, useLayoutEffect, useState } from 'react';
 import { TranslationContext } from './components/TranslationContext';
 import { useConnectionContext } from './components/ConnectionContext';
 import { useLocation } from 'react-router-dom';
@@ -27,6 +27,23 @@ import AdminPassword from './components/AdminPassword';
 import ControlHandler from './components/ControlHandler';
 import iro from '@jaames/iro';
 import ControlWindow from './assets/ControlWindow';
+
+interface Channel {
+  id: number;
+  channel_type: 'main' | 'bi' | 'r' | 'g' | 'b';
+  sliderValue: number;
+}
+
+interface DeviceConfig {
+  id: number;
+  deviceValue: number;
+  name: string;
+  attributes: {
+    channel: Channel[];
+  };
+  device_type: string;
+  universe: string;
+}
 
 // LightFX
 function Control() {
@@ -44,22 +61,14 @@ function Control() {
   const [isSolo, setIsSolo] = useState(false);
   const [height, setHeight] = useState(-3);
   const [selected, setSelected] = useState(selectedDevices[0] && devices.length > 0);
-
-  interface DeviceConfig {
-    id: number;
-    deviceValue: number;
-    name: string;
-    attributes: any;
-    device_type: string;
-    universe: string;
-  }
+  const [deviceModified, setDeviceModified] = useState(false); // Device added or removed
 
   // update selected state
   useLayoutEffect(() => {
     setSelected(selectedDevices[0] && devices.length > 0);
   }, [selectedDevices, devices]);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     const fetchDevices = async (reset: boolean) => {
       try {
         const response = await fetch(url + '/fader');
@@ -144,15 +153,23 @@ function Control() {
     }
   }, [selectedDevices, unselectedDevices, devices]);
 
-  const handleAddDevice = (device: DeviceConfig) => {
-    setSelectedDevices([...selectedDevices, device]);
-    setUnselectedDevices(unselectedDevices.filter((item) => item.id !== device.id));
-  };
+  const handleAddDevice = useCallback(
+    (device: DeviceConfig) => {
+      setSelectedDevices((current) => [...current, device]);
+      setUnselectedDevices((current) => current.filter((item) => item.id !== device.id));
+      setDeviceModified(true);
+    },
+    [selectedDevices, unselectedDevices]
+  );
 
-  const handleRemoveDevice = (device: DeviceConfig) => {
-    setSelectedDevices(selectedDevices.filter((s) => s.id !== device.id));
-    setUnselectedDevices([...unselectedDevices, device]);
-  };
+  const handleRemoveDevice = useCallback(
+    (device: DeviceConfig) => {
+      setSelectedDevices((current) => current.filter((s) => s.id !== device.id));
+      setUnselectedDevices((current) => [...current, device]);
+      setDeviceModified(true);
+    },
+    [selectedDevices, unselectedDevices]
+  );
 
   // Solo Button
   const toggleSolo = () => {
@@ -177,10 +194,32 @@ function Control() {
     setFaderValue(0, 2, Math.min(255, Math.max(0, Math.round(((tempInKelvin - 2200) / 8800) * 255))));
   };
 
-  // Give changed fader values to ControlHandler
-  /* !!! Wird gespammt wenn mehrere clients connected sind !!! */
+  // Give changed fader values to ControlHandler and FaderContext
   useEffect(() => {
     ControlHandler(selectedDevices, faderValues[0].slice(1), emit);
+
+    // set all channels of selectedDevices to the corresponding channel values of faderValues[0]
+    selectedDevices.forEach((device) => {
+      device.attributes.channel.forEach((channel) => {
+        switch (channel.channel_type) {
+          case 'main':
+            setFaderValue(device.id, 0, faderValues[0][1]);
+            break;
+          case 'bi':
+            setFaderValue(device.id, 1, faderValues[0][2]);
+            break;
+          case 'r':
+            setFaderValue(device.id, 1, faderValues[0][3]);
+            break;
+          case 'g':
+            setFaderValue(device.id, 2, faderValues[0][4]);
+            break;
+          case 'b':
+            setFaderValue(device.id, 3, faderValues[0][5]);
+            break;
+        }
+      });
+    });
   }, [faderValues[0][1], faderValues[0][2], faderValues[0][3], faderValues[0][4], faderValues[0][5]]);
 
   const [effects, setEffects] = useState(true);
@@ -202,6 +241,54 @@ function Control() {
 
     if (!effectFound) setEffects(false);
   }, [selectedDevices]);
+
+  // Transfer fader values from selected devices to the control
+  useEffect(() => {
+    const updateFaderValuesForSelectedDevices = () => {
+      if (selectedDevices.length === 0) return;
+
+      let flags = { mainSet: false, biColorSet: false, rgbSet: false };
+
+      selectedDevices.forEach((device) => {
+        device.attributes.channel.forEach((channel) => {
+          switch (channel.channel_type) {
+            case 'main':
+              if (!flags.mainSet) {
+                setFaderValue(0, 1, faderValues[device.id][0]);
+                flags.mainSet = true;
+              }
+              break;
+            case 'bi':
+              if (!flags.biColorSet) {
+                const kelvin = Math.round((faderValues[device.id][1] / 255) * 8800 + 2200); // faderValue to kelvin
+                setFaderValue(0, 3, iro.Color.kelvinToRgb(kelvin).r);
+                setFaderValue(0, 5, iro.Color.kelvinToRgb(kelvin).b);
+                flags.biColorSet = true;
+                flags.rgbSet = true;
+              }
+              break;
+            case 'r':
+            case 'g':
+            case 'b':
+              if (!flags.rgbSet) {
+                setFaderValue(0, 3, faderValues[device.id][1]);
+                setFaderValue(0, 4, faderValues[device.id][2]);
+                setFaderValue(0, 5, faderValues[device.id][3]);
+                flags.rgbSet = true;
+                flags.biColorSet = true;
+              }
+              break;
+          }
+        });
+      });
+    };
+
+    // If a device was added or removed
+    if (deviceModified) {
+      updateFaderValuesForSelectedDevices();
+      setDeviceModified(false);
+    }
+  }, [deviceModified, selectedDevices]);
 
   return (
     <>
@@ -311,10 +398,10 @@ function Control() {
                         <React.Fragment key={slider.id}>
                           {slider.attributes.channel
                             .filter(
-                              (channel: { id: number; channel_type: string }) =>
+                              (channel: Channel) =>
                                 channel.channel_type !== 'main' && channel.channel_type !== 'r' && channel.channel_type !== 'g' && channel.channel_type !== 'b' && channel.channel_type !== 'bi'
                             )
-                            .map((channel: { id: number; channel_type: string }, channelIndex: number, filteredChannels: string | any[]) => {
+                            .map((channel: Channel, channelIndex: number, filteredChannels: Channel[]) => {
                               const isLastFader = index === totalSliders - 1 && channelIndex === filteredChannels.length - 1;
                               return (
                                 <div key={slider.id + '-' + channel.id}>