+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Updated dependencies
 - Added DMX occupancy overview to LightSettings
 - Scenes are "on" when created
+- LightFX
+  - Transfer the values of the first light of its kind to the corresponding faders
 
 ### Fixes